streamlitInteractiveGraph_app/requirements.txt
streamlit
streamlit-agraph
openai
requests
beautifulsoup4
langchain
langchain-community
langchain-openai
python-dotenv
tiktoken
neo4j